[FIX] Corração de login sem a necessidade de redirecionamento para criar uma instância do VueSocketIO.
#169 #190 #186 #155
